refactor: create util dir for controllers test
