+++ src/test/java/controllers/unit/FlightControllerUnitTest.java
@@ -28,8 +28,8 @@ import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
 
-import static controllers.unit.CommonAssertions.assertErrorResponse;
-import static controllers.unit.CommonAssertions.assertSuccessfulResponse;
+import static controllers.unit.util.CommonAssertions.assertErrorResponse;
+import static controllers.unit.util.CommonAssertions.assertSuccessfulResponse;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.ArgumentMatchers.any;
