+++ src/test/java/controllers/unit/util/CommonAssertions.java
@@ -1,4 +1,4 @@
-package controllers.unit;
+package controllers.unit.util;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
